Refactor grammar.js converter into SRP modules
